Moved dlc data structures from commons to core
