+++ app/gui/src/main/scala/org/bitcoins/gui/dlc/DLCTableView.scala
@@ -1,7 +1,7 @@
 package org.bitcoins.gui.dlc
 
-import org.bitcoins.commons.jsonmodels.dlc.{AcceptedDLCStatus, DLCStatus}
-import org.bitcoins.commons.jsonmodels.dlc.DLCStatus._
+import org.bitcoins.core.protocol.dlc.{AcceptedDLCStatus, DLCStatus}
+import org.bitcoins.core.protocol.dlc.DLCStatus._
 import scalafx.beans.property.StringProperty
 import scalafx.geometry.Insets
 import scalafx.scene.control.{ContextMenu, MenuItem, TableColumn, TableView}
