+++ db-commons/src/main/scala/org/bitcoins/db/DbCommonsColumnMappers.scala
@@ -1,7 +1,6 @@
 package org.bitcoins.db
 
-import org.bitcoins.commons.jsonmodels.dlc.DLCMessage.ContractInfo
-import org.bitcoins.commons.jsonmodels.dlc.{DLCState, SigningVersion}
+import org.bitcoins.core.protocol.dlc.DLCMessage.ContractInfo
 import org.bitcoins.commons.jsonmodels.wallet.{
   WalletStateDescriptor,
   WalletStateDescriptorType
@@ -12,6 +11,7 @@ import org.bitcoins.core.currency.{CurrencyUnit, Satoshis}
 import org.bitcoins.core.gcs.FilterType
 import org.bitcoins.core.hd._
 import org.bitcoins.core.number.{Int32, UInt32, UInt64}
+import org.bitcoins.core.protocol.dlc.{DLCState, SigningVersion}
 import org.bitcoins.core.protocol.script.{ScriptPubKey, ScriptWitness}
 import org.bitcoins.core.protocol.tlv._
 import org.bitcoins.core.protocol.transaction.{
