+++ dlc-oracle-test/src/test/scala/org/bitcoins/dlc/oracle/storage/EventDAOTest.scala
@@ -1,9 +1,8 @@
 package org.bitcoins.dlc.oracle.storage
 
 import java.time.Instant
-
-import org.bitcoins.commons.jsonmodels.dlc.SigningVersion
 import org.bitcoins.core.hd.{HDCoinType, HDPurpose}
+import org.bitcoins.core.protocol.dlc.SigningVersion
 import org.bitcoins.core.protocol.tlv.EventDescriptorTLV
 import org.bitcoins.core.util.TimeUtil
 import org.bitcoins.crypto._
