+++ dlc-oracle/src/main/scala/org/bitcoins/dlc/oracle/OracleEvent.scala
@@ -1,9 +1,8 @@
 package org.bitcoins.dlc.oracle
 
 import java.time.Instant
-
-import org.bitcoins.commons.jsonmodels.dlc.SigningVersion
 import org.bitcoins.core.number.UInt32
+import org.bitcoins.core.protocol.dlc.SigningVersion
 import org.bitcoins.core.protocol.tlv._
 import org.bitcoins.crypto._
 import org.bitcoins.dlc.oracle.storage.EventDb
