+++ dlc-oracle/src/main/scala/org/bitcoins/dlc/oracle/storage/EventDb.scala
@@ -1,8 +1,8 @@
 package org.bitcoins.dlc.oracle.storage
 
-import java.time.Instant
+import org.bitcoins.core.protocol.dlc.SigningVersion
 
-import org.bitcoins.commons.jsonmodels.dlc.SigningVersion
+import java.time.Instant
 import org.bitcoins.core.protocol.tlv._
 import org.bitcoins.crypto._
 import org.bitcoins.dlc.oracle.OracleEvent
